+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 development
   - correct summary value for attestations in epoch
   - batch finality updates, reducing memory usage and increasing responsiveness
+  - avoid issue of duplicate column when upgrading f_parent_distance
 
 0.5.1
   - fix missed sync committee information if first block of period is missed